+++ src/Exception.php
@@ -1,0 +1,19 @@
+<?php
+
+/**
+ * This file is part of Carpediem\JSend, a JSend Response object.
+ *
+ * @copyright Carpe Diem. All rights reserved
+ * @license MIT See LICENSE.md at the root of the project for more info
+ */
+
+declare(strict_types=1);
+
+namespace Carpediem\JSend;
+
+/**
+ * A Immutable Value Object Class to represent a JSend object.
+ */
+class Exception extends \Exception
+{
+}